WIP: include fenics interface in published package, add tests on electrochemistry samples including fenics, update fenics interface docs
