+++ .github/workflows/tests_fenics.yml
@@ -1,0 +1,93 @@
+name: Test electrochemistry examples including fenics functionality
+
+on:
+  push:
+    branches:
+      - '**'
+    tags:
+      - '**'
+
+jobs:
+  tests:
+
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v3
+        with:
+          fetch-depth: 0
+
+      - name: install_python
+        run: |
+          sudo apt-get update -qy
+          sudo apt-get install -y python3-dev python3-pip python3-venv libxml2-dev libxslt-dev zlib1g-dev
+          # Upgrade to latest meson and ninja
+          sudo pip install --upgrade meson ninja
+
+      - name: install_fenics
+        run: |
+          sudo apt-get install -y --no-install-recommends software-properties-common
+          sudo add-apt-repository -y ppa:fenics-packages/fenics
+          sudo apt-get update -qy
+          sudo apt-get install -y fenics=1:2019.1.0.3
+          
+          sudo pip install fenics==2019.1.0
+          
+          PYBIND11_VERSION=2.2.4
+          wget -nc --quiet https://github.com/pybind/pybind11/archive/v${PYBIND11_VERSION}.tar.gz
+          tar -xf v${PYBIND11_VERSION}.tar.gz \\
+            && cd pybind11-${PYBIND11_VERSION}
+          mkdir build \\
+            && cd build \\
+            && cmake -DPYBIND11_TEST=off -DCMAKE_INSTALL_PREFIX=/usr -DPYTHON_EXECUTABLE=$(which python) .. \\ 
+            && sudo make install
+    
+          FENICS_VERSION=$(python -c"import ffc; print(ffc.__version__)")
+          MSHR_VERSION=$(echo $FENICS_VERSION | sed 's/.post[0-9]$//')
+          git clone --branch=${FENICS_VERSION} https://bitbucket.org/fenics-project/dolfin
+          git clone --branch=${MSHR_VERSION} https://bitbucket.org/fenics-project/mshr
+          mkdir dolfin/build \\
+            && cd dolfin/build \\ 
+            && cmake -DPYTHON_EXECUTABLE=$(which python) .. \\
+            && sudo make install \\ 
+            && cd ../..
+          mkdir mshr/build \\
+            && cd mshr/build \\
+            && cmake -DPYTHON_EXECUTABLE=$(which python) .. \\
+            && sudo make install \\
+            && cd ../..
+          cd dolfin/python \\
+            && sudo pip install . \\
+            && cd ../..
+          cd mshr/python && mkdir build \\
+            && cd build \\
+            && cmake -DCMAKE_INSTALL_PREFIX=/usr -DPYTHON_EXECUTABLE=$(which python) .. \\
+            && cd .. \\
+            && sudo pip install . \\
+            && cd ../..
+          cat /usr/share/dolfin/dolfin.conf
+
+      - name: install matscipy
+        run: |
+          sudo pip install .
+
+      - name: run examples
+        run: |
+          source /usr/share/dolfin/dolfin.conf
+          cd examples/electrochemistry
+          
+          # check that all sample html files are produced successfully
+          html_files=(./*.html)
+          rm *.html
+          
+          ls -1 *.py | grep -v to_html.py | bash to_html.sh
+          
+          for f in "${html_files[@]}"; do
+            if ! test -f "$f"; then
+              echo "Sample $f creation failed."
+              exit 1
+            fi
+          done
+          
+          # check that parametric sweep sample runs succesfully 
+          # bash pcs_pipeline.sh
