+++ examples/electrochemistry/pnp_batch/cell_1d/potential_sweep/run.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
-./run_std.sh   | tee run_std.log
-./run_robin.sh | tee run_robin.log
-./eval.sh      | tee eval.log
+bash ./run_std.sh   | tee run_std.log
+bash ./run_robin.sh | tee run_robin.log
+bash ./eval.sh      | tee eval.log
