+++ examples/electrochemistry/samples_pnp_c2d.py
@@ -32,8 +32,11 @@
 
 # %%
 # stretching notebook width across whole window
-from IPython.core.display import display, HTML
-display(HTML("<style>.container { width:100% !important; }</style>"))
+try:
+    from IPython.core.display import display, HTML
+    display(HTML("<style>.container { width:100% !important; }</style>"))
+except:
+    pass
 
 # %%
 # basics
@@ -47,7 +50,6 @@ import matplotlib.pyplot as plt
 from scipy import interpolate
 from matscipy.electrochemistry import continuous2discrete
 from matscipy.electrochemistry import get_histogram
-from matscipy.electrochemistry.utility import plot_dist
 
 # %%
 # electrochemistry basics
@@ -65,6 +67,38 @@ from matscipy.electrochemistry import PoissonNernstPlanckSystem
 # 3rd party file output
 import ase
 import ase.io
+
+# %%
+# helper functions
+def get_centers(bins):
+    """Return the center of the provided bins.
+
+    Example:
+    >>> get_centers(bins=np.array([0.0, 1.0, 2.0]))
+    array([ 0.5,  1.5])
+    """
+    bins = bins.astype(float)
+    return (bins[:-1] + bins[1:]) / 2
+
+def plot_dist(histogram, name, reference_distribution=None):
+    """Plot histogram with an optional reference distribution."""
+    hist, bins = histogram
+    width = 1 * (bins[1] - bins[0])
+    centers = get_centers(bins)
+
+    fi, ax = plt.subplots()
+    ax.bar( centers, hist, align='center', width=width, label='Empirical distribution',
+            edgecolor="none")
+
+    if reference_distribution is not None:
+        ref = reference_distribution(centers)
+        ref /= sum(ref)
+        ax.plot(centers, ref, color='red', label='Target distribution')
+
+    plt.title(name)
+    plt.legend()
+    plt.xlabel('Distance ' + name)
+    plt.savefig(name + '.png')
 
 # %%
 # PoissonNernstPlanckSystem makes extensive use of Python's logging module
