+++ examples/electrochemistry/stericify_batch/cell_1d/potential_sweep/run.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
-./run_std.sh   | tee run_std.log
-./run_robin.sh | tee run_robin.log
+bash ./run_std.sh   | tee run_std.log
+bash ./run_robin.sh | tee run_robin.log
