+++ matscipy/electrochemistry/meson.build
@@ -5,7 +5,9 @@ python_sources = [
     '__init__.py',
     'continuous2discrete.py',
     'poisson_boltzmann_distribution.py',
-    'poisson_nernst_planck_solver.py'
+    'poisson_nernst_planck_solver.py',
+    'poisson_nernst_planck_solver_fenics.py',
+    'steric_correction.py'
 ]
 
 # Install pure Python
